requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
aiofiles==23.2.1
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.6
pillow==10.1.0
python-magic==0.4.27
cryptography>=41.0.0
pytz==2023.3
openai>=1.0.0
colorlog==6.8.0
yt-dlp>=2023.11.0
ffmpeg-python>=0.2.0
numpy==1.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
pyrogram>=2.0.106
TGCrypto>=1.2.5

